Bin all freq components for all locations
Not just top 100, but ALL freq components are binned, to arrive at
a fixed binning, and then this binning is used at each location,
along with low-pass and high-pass filters, to get blueprints of
low and high freq components separately, that are comparable with
other locations.

There are two hyperparameters - threshold for low-pass/high-pass in
cycles per day (3 or 4), and number of bins for ALL freqs (100).